Add teaser data adapter pattern in component registry
